# listingcourtscheduler-api/src/main/resources/openapi/courtscheduler-api.openapi.yml
openapi: 3.0.3
info:
  title: Court Scheduler API
  description: |
    Spring Boot port of the existing RAML 0.8 contract under
    listingcourtscheduler-api/src/raml/courtscheduler-api.raml.
    Paths, query parameters, vendor media types and status codes are preserved
    to keep the existing UI clients working without change.
  version: 17.0.76
servers:
  - url: /listingcourtscheduler-api/rest/courtscheduler
    description: Local / WildFly-equivalent context path
...
paths:
...
  /judiciaries/availability-rules:
    get:
      tags: [judiciary-availability]
      operationId: findJudiciaryAvailabilityRules
      parameters:
        - { in: query, name: startDate,    required: true,  schema: { type: string } }
        - { in: query, name: endDate,      required: true,  schema: { type: string } }
        - { in: query, name: courtCentreId, required: false, schema: { type: string } }
        - { in: query, name: pageSize,     required: false, schema: { type: integer, default: 20 } }
        - { in: query, name: pageNumber,   required: false, schema: { type: integer, default: 1 } }
        - { in: query, name: withJudiciary, required: false, schema: { type: boolean, default: true } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourtschedulerFindJudiciaryAvailabilityRule'
  /judiciaries/availability-rules/{ruleId}:
    get:
      tags: [judiciary-availability]
      operationId: getJudiciaryAvailabilityRule
      parameters:
        - in: path
          name: ruleId
          required: true
          schema: { type: string }
        - in: query
          name: withJudiciary
          required: false
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourtschedulerGetJudiciaryAvailabilityRule'
  /judiciaries/availability-rules/add:
    post:
      tags: [judiciary-availability]
      operationId: addJudiciaryAvailabilityRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtschedulerAddJudiciaryAvailabilityRule'
      responses:
        '200':
          description: OK
        '422':
          description: Validation failed
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
  /judiciaries/availability-rules/update:
    post:
      tags: [judiciary-availability]
      operationId: updateJudiciaryAvailabilityRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtschedulerUpdateJudiciaryAvailabilityRule'
      responses:
        '200':
          description: OK
        '422':
          description: Validation failed
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
  /judiciaries/availability-rules/delete:
    post:
      tags: [judiciary-availability]
      operationId: deleteJudiciaryAvailabilityRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtschedulerDeleteJudiciaryAvailabilityRule'
      responses:
        '200':
          description: OK
        '422':
          description: Validation failed
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
  /judiciaries/availability-rules/validate-add:
    post:
      tags: [judiciary-availability]
      operationId: validateAddJudiciaryAvailabilityRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtschedulerAddJudiciaryAvailabilityRule'
      responses:
        '200':
          description: OK
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
        '422':
          description: Validation failed
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
  /judiciaries/availability-rules/validate-update:
    post:
      tags: [judiciary-availability]
      operationId: validateUpdateJudiciaryAvailabilityRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtschedulerUpdateJudiciaryAvailabilityRule'
      responses:
        '200':
          description: OK
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
        '422':
          description: Validation failed
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
  /judiciaries/availability-rules/validate-delete:
    post:
      tags: [judiciary-availability]
      operationId: validateDeleteJudiciaryAvailabilityRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtschedulerDeleteJudiciaryAvailabilityRule'
      responses:
        '200':
          description: OK
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
        '422':
          description: Validation failed
          content:
            application/vnd.courtscheduler.validate.response+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerValidateResponse'
  /judiciaries:
    get:
      tags: [judiciary-availability]
      operationId: findJudiciaryAvailability
      summary: |
        Judiciary queries, selected by response media type:
        search.available.judiciaries — typeahead search over reference-data judiciaries with optional
        in-courtscheduler availability filtering; application/json — judiciary availability over a
        date range (was GET /judiciaries/availability). Query parameters are the union of both
        variants and are therefore all optional at the contract level; each action validates its own
        required parameters.
      parameters:
        - { in: query, name: search,          required: false, schema: { type: string } }
        - { in: query, name: judiciaryGroup,  required: false, schema: { type: string } }
        - { in: query, name: limit,           required: false, schema: { type: string } }
        - { in: query, name: dates,           required: false, schema: { type: string } }
        - { in: query, name: courtHouseId,    required: false, schema: { type: string } }
        - { in: query, name: courtScheduleIds, required: false, schema: { type: string } }
        - { in: query, name: ignoreAvailability, required: false, schema: { type: boolean, default: false } }
        - { in: query, name: startDate,    required: false, schema: { type: string } }
        - { in: query, name: endDate,      required: false, schema: { type: string } }
        - { in: query, name: courtCentreId, required: false, schema: { type: string } }
        - { in: query, name: pageSize,     required: false, schema: { type: integer, default: 20 } }
        - { in: query, name: pageNumber,   required: false, schema: { type: integer, default: 1 } }
        - { in: query, name: withJudiciary, required: false, schema: { type: boolean, default: true } }
      responses:
        '200':
          description: OK
          content:
            application/vnd.courtscheduler.search.available.judiciaries+json:
              schema:
                $ref: '#/components/schemas/CourtschedulerSearchAvailableJudiciaries'
            application/json:
              schema:
                $ref: '#/components/schemas/CourtschedulerFindJudiciaryAvailability'
components:
  schemas:
...
    PassthroughObject:
      type: object
      additionalProperties: true
...
    CourtschedulerSearchAvailableJudiciaries:    { $ref: '#/components/schemas/PassthroughObject' }
...
    CourtschedulerValidateResponse:               { $ref: '#/components/schemas/PassthroughObject' }
    CourtschedulerFindJudiciaryAvailabilityRule:  { $ref: '#/components/schemas/PassthroughObject' }
    CourtschedulerGetJudiciaryAvailabilityRule:   { $ref: '#/components/schemas/PassthroughObject' }
    CourtschedulerAddJudiciaryAvailabilityRule:   { $ref: '#/components/schemas/PassthroughObject' }
    CourtschedulerUpdateJudiciaryAvailabilityRule: { $ref: '#/components/schemas/PassthroughObject' }
    CourtschedulerDeleteJudiciaryAvailabilityRule: { $ref: '#/components/schemas/PassthroughObject' }
    CourtschedulerFindJudiciaryAvailability:      { $ref: '#/components/schemas/PassthroughObject' }